requirements.txt
amqp==5.1.0
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.0
pika==1.2.0
pika-stubs==0.1.3
prompt-toolkit==3.0.29
python-dotenv==0.20.0
pytz==2022.1
pyjwt==2.3.0
six==1.16.0
SQLAlchemy==1.4.34
uwsgidecorators==1.1.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
zipp==3.8.0
uWSGI==2.0.20
